+++ e2e_tests/constants.ts
@@ -1,0 +1,3 @@
+export const API_KEY = process.env.API_KEY!;
+export const OWNER_MNEMONIC = process.env.OWNER_MNEMONIC!;
+export const PAYMASTER_POLICY_ID = process.env.PAYMASTER_POLICY_ID!;
